new changes

+++ python-object_relational_mapping/13-model_state_delete_a.py
@@ -4,35 +4,18 @@ Script that deletes all State objects with a name containing the letter 'a'
 """
 
 import sys
-from model_state import Base, State
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
+from model_state import State
 
 if __name__ == "__main__":
-    # Get command line arguments
-    username = sys.argv[1]
-    password = sys.argv[2]
-    database = sys.argv[3]
-
-    # Create engine to connect to MySQL server
-    engine = create_engine('mysql+mysqldb://{}:{}@localhost:3306/{}'.format(
-        username, password, database), pool_pre_ping=True)
-
-    # Create a configured "Session" class
+    engine = create_engine("mysql+mysqldb://{}:{}@localhost/{}"
+                           .format(sys.argv[1], sys.argv[2], sys.argv[3]),
+                           pool_pre_ping=True)
     Session = sessionmaker(bind=engine)
-
-    # Create a Session instance
     session = Session()
 
-    # Query all State objects that contain the letter 'a' in their name
-    states_to_delete = session.query(State).filter(State.name.contains('a')).all()
-
-    # Delete each state that contains 'a'
-    for state in states_to_delete:
-        session.delete(state)
-
-    # Commit the changes to the database
+    for state in session.query(State):
+        if "a" in state.name:
+            session.delete(state)
     session.commit()
-
-    # Close the session
-    session.close()
